A3 follow-up: widening multiply for the squared-magnitude accumulator
The ESP-IDF component-size report on the PR exposed a new libgcc.a dependency,
187 bytes, absent from the same report on main. The cause is A3's own change:
declaring the scaled magnitude as uint64_t made `scaled * scaled` a full 64x64
multiply, which a 32-bit target implements with a __muldi3 helper call -- once
per source sample.

That is the same class of cost A3 exists to remove. Replacing a soft-float
double multiply with a soft-int64 multiply is not progress.

The scaled magnitude is at most 2^23, so it fits uint32_t, and
`(uint64_t)scaled * scaled` is the widening 32x32->64 pattern that 32-bit ISAs
provide natively. The accumulator stays uint64_t; only the operand type
changes.

Published column values remain bit-identical to the pre-A3 build, verified by
diffing overview output against a worktree at the A2 commit. 72/72 tests pass.

Found only because the PR ran CI: no RISC-V or Xtensa toolchain is available
locally, and on x86-64 both forms compile to a single instruction.

+++ src/waveform/apta_waveform_detail_mvp.c
@@ -289,7 +289,7 @@ apta_status_t apta_internal_detail_process_sample(
     uint32_t tile_index;
     uint32_t column_index;
     float magnitude;
-    uint64_t scaled;
+    uint32_t scaled;
     apta_internal_detail_tile_t *tile;
     apta_internal_waveform_accumulator_t *accumulator;
 
@@ -324,8 +324,9 @@ apta_status_t apta_internal_detail_process_sample(
     }
     /* A3: integer accumulation with a branchless clamp. */
     magnitude = fminf(fabsf(sample), 1.0f);
-    scaled = (uint64_t)(magnitude * APTA_INTERNAL_SQUARE_MAGNITUDE_SCALE);
-    accumulator->sum_squares += scaled * scaled;
+    /* Widening 32x32->64 multiply; see apta_waveform_process.c. */
+    scaled = (uint32_t)(magnitude * APTA_INTERNAL_SQUARE_MAGNITUDE_SCALE);
+    accumulator->sum_squares += (uint64_t)scaled * scaled;
     accumulator->sample_count += 1u;
     if (sample <= -1.0f || sample >= 1.0f) {
         accumulator->clipped = 1u;
